requirements.txt
fastapi
uvicorn
python-dotenv
dspy-ai
anthropic